# fake-rest-client/src/main/resources/swagger.yml
openapi: 3.0.1
info:
  title: FakeRESTApi.Web V1
  version: v1
servers:
  - url: https://fakerestapi.azurewebsites.net/
paths:
...
  /api/v1/Books:
    get:
      operationId: getBooks
      tags:
        - Books
      responses:
        '200':
          description: Success
          content:
            application/json; v=1.0:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      operationId: createBook
      tags:
        - Books
      requestBody:
        content:
          application/json; v=1.0:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Success
  /api/v1/Books/{id}:
    get:
      operationId: getBookById
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json; v=1.0:
              schema:
                $ref: '#/components/schemas/Book'
    put:
      operationId: updateBook
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json; v=1.0:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Success
    delete:
      operationId: deleteBook
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        pageCount:
          type: integer
          format: int32
        excerpt:
          type: string
          nullable: true
        publishDate:
          type: string
          format: date-time
      additionalProperties: false
...